Move TypeScript definitions into billing-models.
