+++ src/billing/BillingAccount/BillingAccountControls.tsx
@@ -4,12 +4,12 @@ import React from 'react';
 import { useState } from 'react';
 import * as Auth from 'src/auth/auth';
 import { accountLinkStyle } from 'src/billing/utils';
+import { GCPBillingProject, GoogleBillingAccount } from 'src/billing-core/models';
 import { VirtualizedSelect } from 'src/components/common';
 import { TextInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import { MenuTrigger } from 'src/components/PopupTrigger';
 import { Ajax } from 'src/libs/ajax';
-import { GCPBillingProject, GoogleBillingAccount } from 'src/libs/ajax/billing/Billing';
 import { reportErrorAndRethrow } from 'src/libs/error';
 import Events, { extractBillingDetails } from 'src/libs/events';
 import { FormLabel } from 'src/libs/forms';
